there was too much to break up....
iin short, we deprecate all mention of credentials, and move to making a config. The rest of the changes handle how that one change affects everything else and putting it into place.

Then number of changes to the the messaging, the unzip hadling, large number of changes resultant in the ui handling.

All the pkgdown as well, and reformatting most tests to handle the change