New vigneette detailing how to run the full test suite locally.
And then a couple of tweaks to the API tests, so that timeouts are handled such that they wil be skipped i tests if the API is being slow. When not in tests, this will also now cause a nice stop message explaining why your API request or similar has failed.